+++ mr-cooper-hackthon/src/main/java/com/hackthon/bo/SentimentalPolarity.java
@@ -2,16 +2,16 @@ package com.hackthon.bo;
 
 public enum SentimentalPolarity {
 
-  POSITIVE("Positive", 0.5f), NEGATIVE("Negative", 0.75f), NEUTRAL("Negetive", 0f);
+  POSITIVE("Positive", 0.5), NEGATIVE("Negative", 0.75), NEUTRAL("Negetive", 0.0);
   
-  private SentimentalPolarity(String value, Float confidenceScoreCutOff) {
+  private SentimentalPolarity(String value, Double confidenceScoreCutOff) {
     this.value = value;
     this.confidenceScoreCutOff = confidenceScoreCutOff;
   }
 
   private String value;
   
-  private Float confidenceScoreCutOff;
+  private Double confidenceScoreCutOff;
   
   public SentimentalPolarity getSentimentalPolarity(String value) {
